+++ examples/src/scenes/CubeScene.h
@@ -11,7 +11,7 @@
 #define CBIT_CUBESCENE_H
 
 #include "core/project/Scene.h"
-#include "../../../src/core/ShaderProgram.h"
+#include "../../../src/core/graphic/ShaderProgram.h"
 #include "../../../src/core/camera/Camera.h"
 
 class CubeScene final : public Scene {
