+++ examples/src/scenes/MeshScene.cpp
@@ -8,7 +8,7 @@
  */
 
 #include "MeshScene.h"
-#include "../../src/core/Locator.h"
+#include "../../../src/core/locator/Locator.h"
 #include "glm/ext/matrix_clip_space.hpp"
 #include "glm/ext/matrix_transform.hpp"
