+++ examples/src/scenes/MeshScene.h
@@ -10,7 +10,7 @@
 #ifndef CBIT_MESHSCENE_H
 #define CBIT_MESHSCENE_H
 
-#include "../../../src/core/MeshQuad.h"
+#include "../../../src/core/mesh/MeshQuad.h"
 #include "core/project/Scene.h"
 
 class MeshScene : public Scene {
