+++ src/core/graphic/TextRenderer.cpp
