+++ src/core/graphic/TextRenderer.h
