+++ src/core/graphic/Texture.cpp
