+++ src/core/mesh/Quad.cpp
