+++ src/core/project/AssetManager.h
